Fix flaky TaskTest.outputDriverFinishEarly
Summary:
TaskTest.outputDriverFinishEarly executes a query with a Values operator, a Limit
operator, and a LocalMerge in between.

The expectation is that the Driver with the Values operator will get called until the
limit is reached and then block.  Then the Driver with the Limit operator will detect
the limit is reached and the Task will shutdown, the goal of the test is to ensure the
Driver is terminated cleanly.

However, since the two Drivers are executed on separate threads, it's possible that
the Driver with the Values operator only gets through one iteration, then the Driver
with the Limit operator executes and completes before the Values operator can
begin its second iteration.  I think this violates the spirit of the test (AFAICT the
author's original intention).

To fix this I added a promise that gets fulfilled when the Values operator reaches
the second iteration, just before it gets blocked.  The Driver with the Limit operator
then gets blocked when noMoreInput is called on the CallbackSink following the
Limit, before it terminates the Task, waiting for a Future of the Promise to be
fulfilled.

This way we guarantee the state I think the author intended to test.

Differential Revision: D46917266

fbshipit-source-id: 5e6ecc8074610485f6e4b23c9ba07b7f67e2c39a