FIX: Refactor computeMetaData and separate propagateNull() from computePropagateNulls() API. (#5287)
Summary:
Pull Request resolved: https://github.com/facebookincubator/velox/pull/5287

This fix a set of expression fuzzer failures such as https://github.com/facebookincubator/velox/issues/5059.

The diff split propagateNull() and computePropagateNulls() API. previously propagateNull()
used to retrieve the propagateNull property for some expressions and re-compute it for others.

with this change propagateNull() only accesses the pre-computed property. And
computePropagateNulls() is called from computeMetaData to compute the property
for some expressions.

The diff also refactor computeMetaData for better readability, it used to interleave computations
of different properties and combine things in a hard to track manner.

Following diffs will address two remaining issues:
- The re-compute that happens over and over during query compilations.
- Avoid emptying distinct fields if shared with parents and simplify the API.

Differential Revision: https://www.internalfb.com/diff/D46457021?entry_point=27

fbshipit-source-id: 3c56b789054ae520391ef9cdc4f6a31c2347f63e