Fix potential concurrency bug in SsdFile::updateStats (#5367)
Summary:
Pull Request resolved: https://github.com/facebookincubator/velox/pull/5367

The test AsyncDataCacheTest.ssd is currently flaky.  When it fails it reports that numPins in the stats from
the SsdCache is more than 0 when it's expected to be 0.

My theory (I'm having trouble consistently reproducing it) is that this is due to cache incoherence between
the threads that are pinning/unpinning regions and the main thread.

The numPins stats is gotten by summing up all the pins across all regions across all SsdFiles in the
SsdCache.  Each SsdFile has an std::vector, counting all the pins across all regions.  Multiple threads update
this std::vector as regions are pinned/unpinned each taking a lock as they do so.  When we call updateStats
on the SsdFile, we don't take a lock (except, somewhat tellingly, in TSAN mode which would complain
otherwise) when we read from that std::vector.

According to the description of the C++ memory model as I understand it, without something to synchronize
threads A and B, if they run on separate cores, B may not see all updates made by A due to cache
incoherence. https://en.cppreference.com/w/cpp/atomic/memory_order (and is also why TSAN is
complaining)

Acquiring a mutex is one way to synchronize threads, so we just need to make that mutex lock happen
across all modes, not just TSAN.

In terms of performance, other than tests
* SsdFile::updateStats() is only called from SsdCache::stats()
* SsdCache::stats() is only called from SsdCache::toString() (which I assume is only used for debugging
purposes) and AsyncDataCache::refreshStats()
* AsyncDataCache::refreshStats() is only called from AsyncDataCache::toString() (which I again assume is
only used for debugging)

So it shouldn't have any impact on production.

Reviewed By: laithsakka

Differential Revision: D46953325

fbshipit-source-id: f9a34beef013692d57188897a84736342eaf57a8