Fix out-of-range condition buffer access in SwitchExpr (#5355)
Summary:
Pull Request resolved: https://github.com/facebookincubator/velox/pull/5355

When evaluating a branch, SwitchExpr first evaluates the condition sub-expression into a `condition` vector. getFlatBool() is then called to extract a  uint64_t* buffer of values in `condition` to participate the subsequent bit operation that computes a SelectivityVector of rows that pass the condition. However, `condition` is guaranteed to be at least of the size `remainingRows->end()` while the bit operaiton is performed from bit 0 to rows.end() that may be larger than remainingRows->end(), hence causing out-of-range access error. This diff fixes this bug by making the bit operation performed from bit 0 to remainingRows->end().

In addition, the `condition` vector was released before its raw values buffer is used in the subsequent bit operation, which is unsafe. This diff also moves the release of the `condition` vector to after the bit operation.

Reviewed By: bikramSingh91

Differential Revision: D46920659

fbshipit-source-id: c172facc06a27e2f97eadb220d505e2bf1a0500f