No need for computeMetaData to be virtual. (#5353)
Summary:
Pull Request resolved: https://github.com/facebookincubator/velox/pull/5353

computeMetaData was virtual, however only one expression namely FieldReference overrides it. The FieldReference computeMetaData calls back the base computeMetaData when there are inputs.
Moreover, metadata for field reference is computed in the same way as other special forms except for distinct fields when there are no inputs.
This diff remove the virtual and handles field reference in computeMetaData

Reviewed By: kagamiori

Differential Revision: D46916994

fbshipit-source-id: 7ab96ac2d0b906873413798696cd704226da0b8a