Fix conjunct clearing pre-existing errors beyond rows.size() (#5365)
Summary:
Pull Request resolved: https://github.com/facebookincubator/velox/pull/5365

Fuzzer found a bug in Conjunct when evaluating an expression
`try(coalesce(map_from_entries(c0), c1, switch(c3, switch((c4) or (c5), c7))))`.
Here `map_from_entiries(c0)` threw at row 99, while `(c4) or (c5)`
was evaluated on a SelectivityVector of size 45. ConjunctExpr calls
finalizeErrors() to clear errors that show up in the initial set of rows
but not in the final set of active rows, which is expected. But what's
wrong is that finalizeErrors() performs bit operation on bit 0--99
(i.e., the size of the pre-existing error vector) with selectivity vectors
of size 45. This leads to out-of-range memory access that either
causes ASAN error or incorrect clear of pre-existing errors. This diff
fixes this problem.

This diff fixes https://github.com/facebookincubator/velox/issues/5364.

Reviewed By: laithsakka

Differential Revision: D46943921

fbshipit-source-id: f9c615d91bd162f00a16ba380297041637e4c9d2