Load vector before serializing (#5140)
Summary:
Pull Request resolved: https://github.com/facebookincubator/velox/pull/5140

For github issue:
https://github.com/facebookincubator/velox/issues/5065

Read the issue for more details about the problem.

If we don't load the lazy vector, we get a SEGV due to asUnchecked().

This approach is hacky. We probably shouldn't pay the cost of checking whether vector is loaded per-row.

For now, it only fails during aggregation functions. So another approach can be to pass ExecCtx to agg functions, and call [OperatorUtils::loadColumns](https://fburl.com/code/gzod4bv5) on the whole vector. Let me know if that would make sense

Reviewed By: xiaoxmeng, Yuhta

Differential Revision: D46415593

fbshipit-source-id: 8a2591521ad76ef8a889df87492d1f59087addb1