Enable huge pages for large contiguous allocations
Hash tables of tens of MB or more benefit from huge pages because of
fewer TLB misses. Savings can be ~30% of CPU time.  We enable huge
pages for allocations over 2MB (huge page is 2MB). Since this is an
advice to the OS that does not change semantics, we do not fail even
if the madvise fails.