adjust simdjson header path (#5415)
Summary:
simdjson does this weird thing in cmake where it copies a header file into the root directory. This prevents us
from using this file in other build systems like buck which does not copy the header file. I modified the cmake
file so that it keeps the header file in the standard location, so that files that import it can be built with cmake
and other build systems.

Pull Request resolved: https://github.com/facebookincubator/velox/pull/5415

Differential Revision: D47041093

Pulled By: kevinwilfong

fbshipit-source-id: 3ee7126fe4f9f33396004cef50e0132707872ece