Avoid resizing child through ensureWritable in FieldReference
Summary: FieldReference used to resize the child vector through BaseVector::ensureWritable when the child size is smaller than its parent size. This resizing was necessary to avoid errors during wrapping, because we wrap the parent's encoding over the child. This leads to increased memory usage in XStream. This diff avoid resizing the child vector by zeroing out indices to NULL rows in the parent's encoding before wrapping it to the child.

Differential Revision: D47045705

fbshipit-source-id: 28f4534e313e0fc515c47ddbc119a503c69647cc