Skip flattening of an expression if inputs are of different types
Existing expression compilation attempt to flatten based on function
names. However, if a function has signatures that do not allow
flattening then this can fail compilation. This is a stopgap
solution that ensures only the most common case is supported where
inputs of the function are of the same type.

Test Plan:
- Added relevant unit tests
- Ran expression fuzzer for 1 hour with assign_function_tickets
startup flag set to 'concat=10'