Size-specific free list in HashStringAllocator
We maintain size specific free lists for faster allocation. When
splitting or merging blocks we may move from one free list to another.
For allocations that do not have to be contiguous, we check the size
specific or larger lists first, then consume from the smaller ones
before growing the HashStringAllocator.

For allocations in the MmapAllocator size classes size range, we go
directly to the allocator of the pool and do not use the
HashStringAllocator arenas. In this way, when such a block is freed it
will become available to the rest of the system, which would not
happen if this were embedded in an extra large HashStringAllocator
arena.

Add a *2 multiplier to needed free space before spilling in
GroupingSet. The test would fail after HashStringAllocator
changes. This is a stopgap for not supporting increasing the
reservation in the spill protected section of GroupingSet.

Fix RowContainer::sizeIncrement(). The min(0,...) would cause any
variable length size increase to be ignored, failing tests after the
HashStringAllocator change. The previous code happened not to fail the
test by dumb luck.