Add support in HiveParser to identify Date logical type
Summary: Date was recently transitions to being a logical type instead of a TypeKind. This meant that HiveParser would translate 'date' hive type to an INTEGER velox type which corresponds to DATE's physical type. However this violated a check in TableHandler which compared velox and hive types. This change ensures that 'date' hive type is correctly converted to it DATE velox type.

Reviewed By: xiaoxmeng, amitkdutta

Differential Revision: D47301882

fbshipit-source-id: f17d1659822675ca81f8bc6d14dd515e54280f96