Make flow control for exchange (#5262)
Summary:
X-link: https://github.com/prestodb/presto/pull/20301

Make flow control for exchange

Adds a request size parameter to ExchangeClient::request(). Adds
counters tracking the history and current state of ExchangeSources,
ExchangeQueue and ExchangeClient.

The flo control of Exchange seeks to work within a maximum queue size
given by, Presto-compatible config name 'exchangeBufferSize()
exchange.max_buffer_size'.

All operations that either consume or process or add sources for
exchange call ExchangeClient::requestIfDue(). This checks if there are
requestable sources and space in the buffer and issues a suitable
number of requests. If this is called from enqueueing a reply from an
exchange source, the source is passed as an argument. requestIfDue
will first request other sources in a round robin fashion. If, after
requesting ther sources, there is space to also request the source
that just arrived, we send a direct request for more. If there is no
space in the queue, we acknowledge the data that may have arrived in
the reply and leave the just arrived source as non-pending. This will
be requested eventually after other sources produce data or time out.

ExchangeQueue tracks the average size of exchange response
pages. Requests are sized to multiples of these. When a request is
started, ExchangeQueue::expectedBytes_ is updated to reflect the
expected data. When a response arrives, the expectedBytes is updated
pro rata, i.e. if there are 3 pending, one response removes at least
1/3 of expectedBytes_.

This mechanism relies on request always producing a possibly empty
response within a relatively short time. This is the case in
Presto/Prestissimo. This is simulated in the test by periodically
timing out all pending callbacks in PartitionedoutputBufferManager.

The test runs exchanges of different widths with different buffer
sizes and variation of data sizes. It checks that the queue peak size
stays within limits.

Requires corresponding changes in Prestissimo PrestoExchangeSource.*.

Pull Request resolved: https://github.com/facebookincubator/velox/pull/5262

Reviewed By: pranjalssh

Differential Revision: D46712319

Pulled By: oerling

fbshipit-source-id: 1c75a07f42a282b608626cb351c2d12127c4625c