Fix unmap sizes
When a mmap is made with a different size and maxsize, the unmap must
unmap maxsize. But the utilization must be decremented by
size. Failure to do this explodes the number of mmaps and fragments
the address space by leaving unused tails of large mmaps hanging after
unmapping the head.

To validate the effect, the test must access the process rss. Does
this with Linux specific ps.