Fix RowContainer::estimateRowSize()
Changes AllocationPool::availableInRun() to mean the space allocatable
without increasing the reservation. Adds availableInAddressRange() to
mean the space in the addresses mapped in the last run of the
AllocationPool. This is the number of allocatable bytes available
before needing a new Allocation/ContiguousAllocation.

Enhance the test for RowContainer::estimateSize(). This test produced
negative sizes before the fix, which went undetected because of
unsigned comparison. This would error out in Deltoid verifier.

Make the test make strings larger than max inline size. Check sizes.