Fix leak of HashStringAllocator external alloc
Destruction/clear() of HashStringAllocator must free allocations it
made directly from its pool(). Normally we do not see this because
uses with aggregates destruct the accumulators before freeing the
HSA. In this way the external allocations are freed when freeing the
accumulator and are not left hanging until the HSA itself is
cleared. But HSA semantics still require free of unfreed external
allocations on destruction since whether an allocation is external or
not is transparent to the user.

Test verifies that clear() frees external allocations. Prior to the
change, only destruction did free these.