Disable Vsize/RSS check
Vsize and RSS reflect actions of the program but ps may not report
these either synchronously or the OS may round up the quantities. So
asserting changes of Vsize or RSS in a test may depend on OS version,
load situation, retrograde Mercury and other factors not controlled by
unit test. The concept of checking that mmap and munmap match is valid
but experience shows this cannot run predictably in an unknown test
environment.