Start to reconcile nodeType and type in SelectiveColumnReader (#5740)
Summary:
Pull Request resolved: https://github.com/facebookincubator/velox/pull/5740

SelectiveColumnReader currently has two types, type and nodeType.  Historically these have been the same,
since we pass in the file type as the requested output type as well as the type in the file.  With the recent changes
to support missing struct fields, these two have diverged. In some readers the nodeType is the requestedType
(mostly primitive types: ints, float points, etc.), while in others it's the file type (complex types like structs, arrays,
and maps).  This leads to bugs when there are missing fields in a struct, because we use the ID from the
nodeType to look up streams/encodings in the file, if nodeType is the requested type, the IDs in the type won't
match the IDs in the file.

To fix this, and hopefully clarify this, I've renamed nodeType and type to requestedType and fileType.  I've also
updated the logic to always use the IDs from fileType.  In general, since we passed fileType for both types in the
past, I generally used fileType for everything.  The exception is in get*Values (getIntValues, getFloatValues, etc.)
where we construct the output Vector, here we want to use requestedType to enable upcasting (the logic is there
to support it, we just didn't have it because we weren't passing in the right type).

I don't think I've cleaned up everything that could be clarified, but this is already a large change.  This is enough
Ato establish requestedType and fileType and fix the bugs referenced above.

Note that Parquet still uses the file type for nodeType/type (now requestedType/fileType) so it's behavior is
unaffected by this change).

Reviewed By: Yuhta

Differential Revision: D47637842

fbshipit-source-id: a2ef83196a3e126a11148b3f9d12c10ffebbb375