ensureLoadedRows should recursively load nested LazyVectors (#5741)
Summary:
Pull Request resolved: https://github.com/facebookincubator/velox/pull/5741

ensureLoadedRows is supposed to ensure that the selected rows are fully loaded.  Unfortunately, today this
doesn't work with nested Lazy Row types, the top level Lazy Row will be loaded, but the nested ones will not.

The fix for this is for ensureLoadedRows to check isLazyNotLoaded on the LazyVector before it returns, and if it
returns false, that means that there is some nested Lazy Vector that needs to be loaded, so it should call itself
recursively.

One tricky point is that because we're decoding Vectors, the SelectivityVector needs to be updated to point to
the corresponding rows in the decoded Vector as we recurse.

Differential Revision: D47638265

fbshipit-source-id: a1f5404e190e27dcb6bc85c7b7f9d29ccff8479a