Add SIMD-based implementation of json_extract (#5535)
Summary:
Pull Request resolved: https://github.com/facebookincubator/velox/pull/5535

Continuing with the work in https://github.com/facebookincubator/velox/pull/5019 and building on
https://github.com/facebookincubator/velox/pull/5525 this adds a version of the JsonExtract UDF with
the underlying implementation swapped from folly to simdjson.

This shows a significant performance improvement, though not as big as the other UDFs with large inputs, in
part due to the fact the large inputs are generated by growing the size of an array, which given the way I
constructed the benchmark, grows the size of the output array proportionately.

Reviewed By: laithsakka

Differential Revision: D47249161

fbshipit-source-id: 49ea0b2289447735a6d715ee0c31beee0c87a9bb