Disable checking window aggregations results with duckdb results. (#5750)
Summary:
Pull Request resolved: https://github.com/facebookincubator/velox/pull/5750

DuckDB has a bug in its window aggregations results.
https://github.com/facebookincubator/velox/pull/5082

We keep seeing failures from time to time in our nightly jobs,
its not worth spending a lot of time investigating each if there is a good
probability its a DuckDB issue.

Furthermore, it sounds like global aggregation significantly also cover parts of
the testing done for window aggregations.

We keep those running for now and just disable the comparison with DuckDB
we will re-enable it once the problem is fixed on DuckDB side.

If you are interested to learn more about the issue dive through the links
starting from those issues

https://github.com/facebookincubator/velox/issues/4350
https://github.com/facebookincubator/velox/issues/4350

Differential Revision: D47645849

fbshipit-source-id: ed35718e5a691430b5348cccb3e3d3bded72c582