Avoid rethrow in setError (#5615)
Summary:
Pull Request resolved: https://github.com/facebookincubator/velox/pull/5615

applyToSelectedpNoThrow catches the thrown exception if any is thrown in func()
When the thrown exception is a Velox exception, it will be thrown again inside setError
when  toVeloxException is called. However since when we already know the exception
type is VeloxException we do not need to rethrow.

This diff introduces context.setVeloxExceptionError to avoid that.

runtime for cast benchmark cast_int##try_invalid_empty_input
  655.78ms  -> 465.23ms

Reviewed By: kevinwilfong

Differential Revision: D47351850

fbshipit-source-id: 4d1817c3578ebfad14cee1846471fe749bca147a