Fix aggregate merge containers and chekc RowContainer variable length free
In the case of merging after spilling, we have a source RowContainer
and a merge RowContainer that share the same Aggregates. Therefore
they must also share the same HashStringAllocator since this is
referenced from the Aggregates. Adds a constructor parameter for
sharing the stringAllocator of another RowContainer.

We add a flag to check that deleting the rows from a RowContainer
causes its stringAllocator to go empty. This is needed in order to
enforce that aggregates, when spilling, free the space they should in
the container.