Fix GroupingSet::toIntermediate fast path
Summary:
When multiple aggregates exist and only some of them support
`toIntermediate`, we still need to construct the `RowContainer` for those who do
not support `toIntermediate`.  In this case when we clean up the `RowContainer`,
we try to destroy the rows using the aggregates with fast path by mistake and
result in bad memory access.  Fix this by skipping calling destroy for aggregate
with fast path in `RowContainer::eraseRows`.

Differential Revision: D47709835

fbshipit-source-id: 59e0466bbc4ea5eca69667220cfd09a9e82ae5de