Track the file ages for AyncDataCache and SsdCache
Use the file open time to calculate the age of AsyncDataCache
and SsdCache entries that are loaded from the file. Tracking this
enables us to expose the metrics maxAge possible of cache entries.

Add a process-wide singleton CacheTTLController to handle the TTL
of AsyncDataCache and SsdCache and track the required info. It
currently adds the file open time when a new file handle is opened. It
also calculates the maxAge for cache entries.