Track entry age for AyncDataCache and SsdCache
Use the raw file open time of the source raw files to calculate
entry ages of the AsyncDataCache and the SsdCache. Track the min, max
and average entry ages.

Add a process-wide FileInfoMap to track the file info for every
file handle opened. Currently tracked file info is file open time
and whether or not the file is in AsyncDataCache or SsdCache.