Introduce compressionKind in HiveInsertTableHandle (#5795)
Summary:
As of now, velox use config to decide compression type for write.
We want the ability to use the dynamically passed-in compression
type by application.

In this PR, we introduce compressionKind in HiveInsertTableHandle
to give applications the ability to do so, if it's not passed in,
we use the config compressionKind by default.

Tests on parquet writer using compression would be added once
https://github.com/facebookincubator/velox/issues/5796  and https://github.com/facebookincubator/velox/issues/5865 are resolved

Pull Request resolved: https://github.com/facebookincubator/velox/pull/5795

Reviewed By: xiaoxmeng

Differential Revision: D47819287

Pulled By: kewang1024

fbshipit-source-id: de944306339eef12766c838d3bb27efff3d3c7a9