Make field not found error a velox error and check valid construction of fieldReference (#5892)
Summary:
Pull Request resolved: https://github.com/facebookincubator/velox/pull/5892

Field not found error ideally should be caught during compilation before the expression runs.
if it was encountered when the expression is evaluated it should not be something
that is hidden by try, hence its type should not be veloxUserError but rather
VeloxError. Since VeloxUserErrors are those that are expected to be hidden by a
try.
This adds checks to catch this issue asap, and change the error type to VeloxError.
A follow up diff should assert that we catch such issue for input row field
referenced during query compilations if do not do that already.

Differential Revision: D47853552

fbshipit-source-id: 2bd6a510e5463869b7c22ff9e13c046de1e25d59