Allow freeing rowsacross many RowContainers
In freeing rows in RowContainer::eraseRows() allow the RowContainer to
filter out the rows that are not allocated from it.  This does ont
occur in usual cases in production but does happen in tests where
tables composed from multiple containers have a fraction of the data
freed.

We add a function to filter out the members of a given container from
a set of arbitrary pointers. We optionally apply this in
RowContainer::eraseRows().