Check column selector against requested type in struct reader (#5927)
Summary:
Pull Request resolved: https://github.com/facebookincubator/velox/pull/5927

The `ColumnSelector` is constructed using schema type not the file
type, so when we check against it, we also need to use schema type.  This was
causing assertion failure when a struct column in file has more fields than the
column in schema (i.e. we drop some fields during schema evolution).

Reviewed By: oerling

Differential Revision: D47897341

fbshipit-source-id: 08192090f3a4baa62087327f67e883e0d00b9cda