Fix ma os CircleCI builds
Currently mac os circle Ci builds are failing as they do not have ccache
installed. The installation script is not executed if a previous cache
with installed dependencies exists. Ideally ccache should be there in
the cache but its not, so we always run the setup script incase
something fails to be included in the cache.