Add fast map for assigning ids to int64 values (#5838)
Summary:
Adds a BigintIdMap utility for use in mapping 64 bit values to unique ids. Can be used in writing dictionary encoding or in VectorHasher. This is 2-4x faster than the corresponding use of F14FastSet.

The table consists of 12 byte entries with 64 bits of value and 32 bits of id. We process 4 values at a time with avx2.

There is a bit mask to disable some values, e.g. for nulls in VectorHasher. These come out as 0. The ids assigned to values start at 1 to distinguish from 0 for non-active lane. The value 0 is used to mark an empty slot in the table. Hence an input of 0 must be treated separately.

The test prints out clock counts for different sizes and frequencies of repeats. The test also checks special cases with inactive lanes and zeros.

Pull Request resolved: https://github.com/facebookincubator/velox/pull/5838

Reviewed By: Yuhta

Differential Revision: D47787041

Pulled By: oerling

fbshipit-source-id: 3c9688445a2369f2a110a0ba5676af2a5f285750